Connected auth BE with FE

+++ backend/routes/posts.js
@@ -31,7 +31,8 @@ router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath:  url + '/images/' + req.file.filename
+    imagePath:  url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -58,12 +59,18 @@ router.put('/:id', checkAuth, multer({storage: storage}).single('image'), (req,r
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId // note that that req has null value on this field cause of security.
   });
 
-  Post.updateOne({_id: req.params.id}, post).then( result => {
-    console.log('Sucess update on: ' + post);
-    res.status(200).json({message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then( result => {
+    if (result.nModified > 0) {
+      console.log('Sucess update on: ' + post);
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      console.log('Not authorized update on: ' + post);
+      res.status(401).json({ message: 'Not authorized!' })
+    }
   })
 });
 
@@ -102,10 +109,15 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      console.log('Sucessful delete on: ');
+      res.status(200).json({message: 'Post Deleted!'});
+    } else {
+      console.log('Not authorized delete!');
+      res.status(401).json({ message: 'Not authorized!' })
+    }
   });
-  res.status(200).json({message: 'Post Deleted!'});
 });
 
 module.exports = router;
